feat: put sidebar seperate
